requirements.txt
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-boost==1.7.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
Pygments==2.7.3
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.2
user-agents==2.2.0
django-heroku
